app/requirements.txt
websockets==11.0.3
python-dotenv==1.0.1
boto3==1.34.0
openai==1.12.0
